requirements.txt
djangorestframework>=3.0,<4.0
psycopg2-binary>=2.8
dj-rest-auth[with_social]>=2.1.3
djangorestframework-simplejwt>=4.6
celery[redis]>=5.0.5
django-model-utils>=4.1.1
